databene-commons-core/src/test/java/org/databene/commons/find/ExtremeElementFinderTest.java
/*
 * (c) Copyright 2007-2009 by Volker Bergmann. All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, is permitted under the terms of the
 * GNU General Public License.
 *
 * For redistributing this software or a derivative work under a license other
 * than the GPL-compatible Free Software License as defined by the Free
 * Software Foundation or approved by OSI, you must first obtain a commercial
 * license to this software product from Volker Bergmann.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 * WITHOUT A WARRANTY OF ANY KIND. ALL EXPRESS OR IMPLIED CONDITIONS,
 * REPRESENTATIONS AND WARRANTIES, INCLUDING ANY IMPLIED WARRANTY OF
 * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE OR NON-INFRINGEMENT, ARE
 * HEREBY EXCLUDED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
 * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
 * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
 * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
 * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
 * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 * POSSIBILITY OF SUCH DAMAGE.
 */

package org.databene.commons.find;

import org.junit.Test;
import static org.junit.Assert.*;

import java.util.Collection;
import java.util.Arrays;

import org.databene.commons.Element;
import org.databene.commons.visitor.CollectionElement;
import org.databene.commons.visitor.ExtremeElementFinder;

/**
 * Tests the {@link ExtremeElementFinder}.<br/><br/>
 * Created: 04.02.2007 09:33:01
 * @author Volker Bergmann
 */
public class ExtremeElementFinderTest {

	@Test
    public void test() {
        Collection<String> collection = Arrays.asList("Alpha", "Bravo");
        Element<String> wrapper = new CollectionElement<String>(collection);
        String min = ExtremeElementFinder.findMin(wrapper);
        assertEquals("Alpha", min);
        assertEquals("Bravo", ExtremeElementFinder.findMax(wrapper));
    }
	
}
